requirements.txt
docxtpl
pony
cvsslib
Flask-WTF
Flask-Limiter
jinja2
flask
uuid
python-magic
mistletoe
aenum